front: rename usePaths -> formatSpaceTimeCurves
