+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/helpers/formatSpaceTimeCurves.ts
